+++ packages/article-byline/src/article-byline-with-links.js
@@ -31,7 +31,8 @@ const ArticleBylineWithLinks = ({ ast, centered, ...props }) =>
   renderByline(
     renderAuthorComponent,
     ast,
-    centered && Platform.OS !== "web"
+    // TODO: revert platform switch after design signoff
+    centered && Platform.OS === "web"
       ? [styles.text, styles.centered]
       : styles.text,
     props