+++ packages/article-main-standard/src/styles/article-header/index.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { fontSizes, spacing } from "@times-components/styleguide";
+import { spacing } from "@times-components/styleguide";
 import globalStyle from "../shared";
 import sharedStyles from "./shared";
 
@@ -7,10 +7,7 @@ const nativeStyles = {
   ...sharedStyles,
   articleHeadLineText: {
     ...sharedStyles.articleHeadLineText,
-    fontSize: fontSizes.headline,
-    fontWeight: "400",
-    lineHeight: 33,
-    marginBottom: spacing(2),
+    fontWeight: "700",
     marginTop: spacing(1)
   },
   articleLabelWrapper: {
@@ -19,9 +16,7 @@ const nativeStyles = {
   },
   standFirst: {
     ...sharedStyles.standFirst,
-    lineHeight: 28,
-    marginTop: -2,
-    paddingBottom: 12
+    lineHeight: 25
   },
   standFirstWithoutFlags: {
     paddingBottom: spacing(4)