+++ packages/article-paragraph/src/styles/index.js
@@ -1,13 +1,13 @@
+import { spacing } from "@times-components/styleguide";
 import sharedStylesFactory from "./shared";
 
 export default (dropCapFont, scale) => {
   const sharedStyles = sharedStylesFactory(dropCapFont, scale);
-
   return {
     ...sharedStyles,
     articleTextElement: {
       ...sharedStyles.articleTextElement,
-      fontStyle: "normal"
+      marginBottom: spacing(4)
     }
   };
 };