+++ packages/provider-queries/src/save-bookmark.js
@@ -1,10 +1,4 @@
 import { addTypenameToDocument } from "apollo-utilities";
-import gql from "graphql-tag";
+import saveBookmarks from "./save_bookmark.graphql";
 
-export default addTypenameToDocument(gql`
-  mutation SaveBookmark($id: UUID!) {
-    saveBookmarks(bookmarks: [{ id: $id }]) {
-      id
-    }
-  }
-`);
+export default addTypenameToDocument(saveBookmarks);