+++ packages/provider-queries/src/topic.js
@@ -1,11 +1,4 @@
 import { addTypenameToDocument } from "apollo-utilities";
-import gql from "graphql-tag";
+import topic from "./topic.graphql";
 
-export default addTypenameToDocument(gql`
-  query TopicQuery($slug: Slug!) {
-    topic(slug: $slug) {
-      name
-      description
-    }
-  }
-`);
+export default addTypenameToDocument(topic);