+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.2.50](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.2.49...@times-components/article-list@9.2.50) (2019-08-28)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.2.49](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.2.48...@times-components/article-list@9.2.49) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.2.48](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.2.47...@times-components/article-list@9.2.48) (2019-08-23)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.2.47](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.2.46...@times-components/article-list@9.2.47) (2019-08-22)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.2.46](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.2.45...@times-components/article-list@9.2.46) (2019-08-21)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.2.45](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.2.44...@times-components/article-list@9.2.45) (2019-08-20)
 
 **Note:** Version bump only for package @times-components/article-list