+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.88](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.2.87...@times-components/error-view@2.2.88) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.2.87](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.2.86...@times-components/error-view@2.2.87) (2019-08-09)
 
 **Note:** Version bump only for package @times-components/error-view