+++ packages/watermark/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.40](https://github.com/newsuk/times-components/compare/@times-components/watermark@2.6.39...@times-components/watermark@2.6.40) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/watermark
+
+
+
+
+
 ## [2.6.39](https://github.com/newsuk/times-components/compare/@times-components/watermark@2.6.38...@times-components/watermark@2.6.39) (2019-08-09)
 
 **Note:** Version bump only for package @times-components/watermark